Enhance error handling and implement ping command
This commit improves error handling by logging the call stack in case of any uncaught exceptions. In 'handleCommands.js' and 'sendEmbed.js', Logger now prints full stack trace of errors, which will help to identify the source of errors more precisely. In 'sendEmbed.js', additional parameters 'ephemeral' and 'components' were added to configure how an Embed is sent to Discord, and error handling was extended to check if 'components' is an array.

Additionally, a new 'ping' command was added. This command responds with bot's latency and other statistics, and includes buttons for users to visit the bot's GitHub page and inviting it to a server.