Add commands and money handling functionalities
A new command loading system has been implemented in handleCommands.js. Commands can now be placed inside directories within the commands folder, making the filesystem cleaner and the commands easier to manage. The commands are then loaded into the bot from these directories.

CreateModel.js file was added to simplify creation of new database models for further MongoDB collections. This avoids having to rewrite common parts of model creation.

Check for bot user and database presence was implemented in the function handling command interactions to prevent bot from interacting with itself and to ensure user and guild data is in the database. These checks ensure that the command is executed with the correct permissions and member information, and improves error handling.

The money.js helper file was added to aid in money handling functionalities for the bot. This consists of functions to add, remove, get, and format money, providing a useful toolkit for creating economy-based commands.