Add logging for channel and cosmetic events
Implemented the logging functionality for channel and cosmetic changes across the application. Added utility functions for fetching log channels, getting channel type names, and responsible user for the events.

This change provides better visibility into channel creation, deletion, and update events. In addition, it lets us track when emojis and stickers are created, deleted, or updated. This information is crucial in moderating and debugging activities in the application.

Signed-off-by: BeauTheBeau <beauthebeau@proton.me>