Refactor slashCommands.js for efficient user and guild creation
Reorganized the 'slashCommands.js' file to efficiently handle the creation of user and guild when not found in the database. Two new helper functions 'createUserIfNotFound' and 'createGuildIfNotFound' were introduced to handle these creations which makes code maintainable. Also, simplified code for embedding and logging messages, to better manage interaction with commands. This provides a better user experience and improves the maintainability of the code.