Update logging and input functionality; enhance commit check presentation
Refactored logging utility to simplify the creation of the log folder and ignore trivial 'Hello, world!' messages. Improved the readability of console prompts in input utility by using the chalk library to add color styling. Revised the commit checking logic in index.js for better code efficiency and added a console message to inform the user when the system is already up-to-date. All changes were made to enhance the user experience and improve the effectiveness of the code.