Add ownerId field and eventListeners command
A new field, OWNER_ID, was added to the environment variables to keep track of the owner of the bot. This is particularly useful for controlling permissions. Additionally, a new 'eventListeners' command was introduced in 'src/commands/dev' which provides a list of all registered event listeners. This command is restricted to the bot owner defined by the OWNER_ID in the.env file.