Add user validation check to memberEventUtility
We added the function "checkUser" from the "makerSurerExister" utility to validate the existence of a user before attempting to access their data in the "memberEventUtility". This was done in order to handle situations where a user may have been deleted or simply does not exist in the database, avoiding potential errors when trying to access non-existing user's data and the join/leave message therefore not being sent.