Update cooldown handling in cooldownManager.js
The logic in the getCooldown function for handling non-existing cooldowns has been improved. Instead of checking existence of cooldowns sequentially and returning false instantly when condition is not met, we now check all conditions in one compound if-statement and only delete the cooldown and return false if the cooldown has already expired. If the required cooldown does not exist, the function now returns resolved Promise for consistent use and behavior across the application.

This prevents users constantly being on cooldown as previously the function was checking if an entry actually existed, if so it returned false. The entry would never be deleted.