Update guildModel with additional fields and categories
Extended the guildSchema in guildModel by adding several new fields such as `notes`, `descriptions`, `protected` and `categories`. In addition, categories were added to each field which will be used to show the field's category in the help command. This was done to enhance the manageability and organization of the guild schema, and to provide more context about each field to the end user.